% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{ATrack}
\alias{ATrack}
\alias{ATrack-class}
\title{ATrack class}
\usage{
ATrack(range = NULL, start = NULL, end = NULL, width = NULL, group, id,
  stacking = "squish", name = "ATrack", fun, selectFun, ...)
}
\arguments{
\item{range,start,end,width,group,id,stacking,name,fun,selectFun,...}{Arguments to be passed to \code{AnnotationTrack}.}
}
\description{
This class contains Gviz's AnnotationTrack and adds default values to the genome and
chromosome slot
}
\examples{
# Object construction
aTrack <- ATrack(start = c(20, 60), end = c(40, 100), name = "random.anno",
id=c("small","big"))
#Stacking example
a2Track <- ATrack(start = c(20, 30), end = c(40, 100), name = "stacking=dense",
id = c("small", "big"), stacking = "dense", fill=c("black", "orange"))
a3Track <- ATrack(start = c(20, 30), end = c(40, 100), name = "no stacking",
id = c("small", "big"), fill = c("black", "orange"))
#Plotting
plotTracks(trackList = c(aTrack, a2Track, a3Track), showFeatureId = TRUE)
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{AnnotationTrack}}, \code{\link{GdObject}}
}

