\name{Pviz_vignette}
\alias{Pviz_vignette}
\alias{seqEx}
\alias{pepEx}
\alias{freqEx}
\alias{posEx}

\title{
Data set for the vignette.
}

\description{
Some sample data sets used for the illustrative examples and the vignette.
\describe{
\item{seqEx}{:	Amino-acid sequence of the envelope of hxb2.}
\item{pepEx}{:	Peptides from peptide microarrays.}
\item{freqEx}{: The expression frequency of each peptide in all analysed the slides.}
\item{posEx}{:	The position of each peptide along the hxb2 reference.}
}
}

\keyword{datasets}
