% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{DTrack}
\alias{DTrack}
\alias{DTrack-class}
\title{DTrack class}
\usage{
DTrack(range = NULL, start = NULL, end = NULL, width = NULL, data,
  name = "DTrack", ...)
}
\arguments{
\item{range,start,end,width,data,name,...}{Arguments to be passed to
\code{DataTrack}.}
}
\description{
This class contains Gviz's DataTrack and adds default values to the genome and
chromosome slot
}
\details{
Refer to \code{DataTrack} for details regarding the constructor.
}
\examples{
dTrack <- DTrack(start=seq(1,1000, len=100), width=10, data=matrix(runif(400),
 nrow=4), name="random data")
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{DataTrack}}, \code{\link{GdObject}}
}

