% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{ProteinAxisTrack}
\alias{ProteinAxisTrack}
\alias{ProteinAxisTrack-class}
\title{ProteinAxisTrack}
\usage{
ProteinAxisTrack(range = NULL, name = "Axis", addNC = FALSE, id = NULL,
  ...)
}
\arguments{
\item{range,name,id,...}{Arguments to be passed to \code{GenomeAxisTrack}.}

\item{addNC}{A \code{logical}. If TRUE, display the Amino-terminal and
  Carboxyl-terminal ends on the axis.}
}
\description{
A track to display an axis for protein or peptide sequences
}
\examples{
# Object construction
paxTrack <- ProteinAxisTrack()
pax2 <- ProteinAxisTrack(addNC=TRUE)
pax3 <- ProteinAxisTrack(littleTicks=TRUE)
# Plotting
plotTracks(c(paxTrack,pax2,pax3), from=1, to=100)
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{GenomeAxisTrack}}
}

