% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{ProteinSequenceTrack}
\alias{ProteinSequenceTrack}
\alias{ProteinSequenceTrack-class}
\title{ProteinSequenceTrack}
\usage{
ProteinSequenceTrack(sequence = NULL, name = "Sequence", ...)
}
\arguments{
\item{sequence}{A \code{character} or \code{AAString} of length one. The
sequence to display.}

\item{name}{A \code{character}. The name of the track used in the title panel
when plotting}

\item{...}{Additional items which will all be interpreted as display
parameters.}
}
\description{
A track to display peptides and protein sequences.
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{SequenceTrack}}, \code{\link{DisplayPars}}
}

