\name{Pviz_vignette}
\alias{Pviz_vignette}
\alias{Ex_seq}
\alias{Ex_pep}
\alias{Ex_freq}
\alias{Ex_pos}

\title{
Data set for the vignette.
}

\description{
Some sample data sets used for the illustrative examples and the vignette.
\describe{
\item{Ex_seq}{:	Amino-acid sequence of the envelope of hxb2.}
\item{Ex_pep}{:	Peptides from peptide microarrays.}
\item{Ex_freq}{: The expression frequency of each peptide in all analysed the slides.}
\item{Ex_pos}{:	The position of each peptide along the hxb2 reference.}
}
}

\keyword{datasets}
